Master

+++ db/migrate/20120425161928_drop_really_old_unused_columns2.rb
@@ -1,15 +1,11 @@
-class DropReallyOldUnusedColumns2 < ActiveRecord::Migration
+class DropReallyOldUnusedColumns2 < ActiveRecord::Migration[4.2]
   tag :postdeploy
 
-  self.transactional = false
+  disable_ddl_transaction!
 
   # cleanup for some legacy database schema that may not even exist for databases created post-OSS release
   def self.maybe_drop(table, column)
-    begin
-      remove_column(table, column)
-    rescue
-      # exception is db-dependent
-    end
+    remove_column(table, column) if self.connection.columns(table).map(&:name).include?(column.to_s)
   end
 
   def self.up
